+++ eprint3x.go
@@ -1067,6 +1067,7 @@ func (e *EPrint) SyntheticFields() {
 				obj["mime_type"] = doc.MimeType
 				obj["content"] = doc.Content
 				obj["license"] = doc.License
+				obj["version"] = fmt.Sprintf("v%d.0.0", doc.RevNumber)
 				if (doc.Placement == 1 || doc.Pos == 1) && doc.Content != "supplemental" {
 					e.PrimaryObject = obj
 				} else {